Entregables/reynaldo_lopez_rubio/APPIUM/requirements.txt
behave==1.2.6
Appium-Python-Client==1.0.2
python-dotenv~=0.19.2
selenium==3.141.0
